Edit create yarn/needle for project
Changed the endpoints for creating a yarn or needle for a specific project so it now works. Also added an endpoint to delete all yarn/needle for a specific project. Added checks for if a yarn/needle exists before putting/patching so a new item is not created if not already in database.