Setup layout for SpecificProject.vue
Made slight changes to Projects.vue so there is now a button that links to a specific project and makes it possible to check how the page looks for testing. Made progress with setting up view for an individual project to display the project information and enable to the user to edit the information.