Change needle/yarn creation forms
Changed the input attributes on the needle & yarn creation forms so the input now reflects what the endpoints expect. Forms now actually submit new yarns and needles. Added a / at the start of the navbar routes so they should no longer go to, for example, specificproject/<numbers here>/project if you click on the project link after viewing a specific project. Changed the methods for creating yarns and needles to the correct endpoint link expected by the API.